+++ karma.conf.js
@@ -12,7 +12,7 @@ module.exports = function(config) {
 
     // frameworks to use
     // available frameworks: https://npmjs.org/browse/keyword/karma-adapter
-    frameworks: ['mocha', 'chai-as-promised', 'chai', 'sinon-chai'],
+    frameworks: ['mocha', 'chai-as-promised', 'chai-dom', 'chai', 'sinon-chai'],
 
   
     client: {